Chopping an image into 170 pieces
Added a function "chop" to ImageToVector.hs that take a width, a height,
and an image and chops the image into as many width by height smaller
images as it can.

Also added preconvert.hs, the compiles to a binary that
will (eventually) take a directory of images, chops each one, and saves
the smaller images in subdirectories.